string_util move remove_post_fix from metadata_util
